% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_cbs_shapefile}
\alias{download_cbs_shapefile}
\title{Download CBS Wijk en Buurtkaart}
\usage{
download_cbs_shapefile(year, path = NULL, verbose = TRUE)
}
\arguments{
\item{year}{Read the shape file from this year.}

\item{path}{Local path to the shape files. If NULL, the shapefiles are
downloaded and a temporary path is created.}

\item{verbose}{Verbosity. Default TRUE.}
}
\value{
The folder to which the files are downloaded.
}
\description{
Download CBS Wijk en Buurtkaart
}
\examples{
download_cbs_shapefile(2017)
download_cbs_shapefile(2017, "data/")
}
