% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{list_cbs_shapefiles}
\alias{list_cbs_shapefiles}
\title{List available CBS wijk en buurtkaarten}
\usage{
list_cbs_shapefiles()
}
\value{
List of available shapefiles
}
\description{
List available CBS wijk en buurtkaarten
}
\examples{
list_cbs_shapefiles()
}
