% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{st_read_cbs}
\alias{st_read_cbs}
\title{Read CBS Wijk en Buurtkaart}
\usage{
st_read_cbs(year, path = NULL, verbose = TRUE)
}
\arguments{
\item{year}{Read the shape file from this year.}

\item{path}{Local path to the shape files. If NULL, the shapefiles are
downloaded and a temporary path is created.}

\item{verbose}{Verbosity. Default TRUE.}
}
\value{
st object with the CBS Wijk en Buurtkaart
}
\description{
Read shapefiles from the CBS Wijk en Buurtkaart with the package sf.
}
\examples{
st_read_cbs(2017)
}
